+++ src/client/components/sidebar.tsx
@@ -9,7 +9,7 @@ import { WorkspaceTree } from "../workspaceTree"
 
 type Props = {
 	mutator: Mutator
-	pageController: IPageController
+    showBoard: (id: string) => void
 	workspaceTree: WorkspaceTree,
 	boardTree?: BoardTree
 }
@@ -47,7 +47,7 @@ class Sidebar extends React.Component<Props> {
 	}
 
 	private showOptions(e: React.MouseEvent, board: Board) {
-		const { mutator, pageController, workspaceTree } = this.props
+		const { mutator, showBoard, workspaceTree } = this.props
 		const { boards } = workspaceTree
 
 		const options: MenuOption[] = []
@@ -64,8 +64,8 @@ class Sidebar extends React.Component<Props> {
 					mutator.deleteBlock(
 						board,
 						"delete block",
-						async () => { pageController.showBoard(nextBoardId!) },
-						async () => { pageController.showBoard(board.id) },
+						async () => { showBoard(nextBoardId!) },
+						async () => { showBoard(board.id) },
 					)
 					break
 				}
@@ -104,20 +104,19 @@ class Sidebar extends React.Component<Props> {
 	}
 
 	private boardClicked(board: Board) {
-		const { pageController } = this.props
-		pageController.showBoard(board.id)
+		this.props.showBoard(board.id)
 	}
 
 	async addBoardClicked() {
-		const { mutator, boardTree, pageController } = this.props
+		const { mutator, boardTree, showBoard } = this.props
 
 		const oldBoardId = boardTree?.board?.id
 		const board = new Board()
 		await mutator.insertBlock(
 			board,
 			"add board",
-			async () => { pageController.showBoard(board.id) },
-			async () => { if (oldBoardId) { pageController.showBoard(oldBoardId) } })
+			async () => { showBoard(board.id) },
+			async () => { if (oldBoardId) { showBoard(oldBoardId) } })
 
 		await mutator.insertBlock(board)
 	}
