+++ src/client/components/tableComponent.tsx
@@ -7,7 +7,7 @@ import { BoardTree } from "../boardTree"
 import { CsvExporter } from "../csvExporter"
 import { Menu } from "../menu"
 import { Mutator } from "../mutator"
-import { IBlock, IPageController } from "../octoTypes"
+import { IBlock } from "../octoTypes"
 import { OctoUtils } from "../octoUtils"
 import { Utils } from "../utils"
 import Button from "./button"
@@ -17,7 +17,10 @@ import { TableRow } from "./tableRow"
 type Props = {
 	mutator: Mutator,
 	boardTree?: BoardTree
-	pageController: IPageController
+    showView: (id: string) => void
+    showCard: (card: IBlock) => void
+    showFilter: (el: HTMLElement) => void
+    setSearchText: (text: string) => void
 }
 
 type State = {
@@ -43,7 +46,7 @@ class TableComponent extends React.Component<Props, State> {
 	}
 
 	render() {
-		const { mutator, boardTree, pageController } = this.props
+		const { mutator, boardTree, showView } = this.props
 
 		if (!boardTree || !boardTree.board) {
 			return (
@@ -85,7 +88,7 @@ class TableComponent extends React.Component<Props, State> {
 					<div className="octo-table">
 						<div className="octo-controls">
 							<Editable style={{ color: "#000000", fontWeight: 600 }} text={activeView.title} placeholderText="Untitled View" onChanged={(text) => { mutator.changeTitle(activeView, text) }} />
-							<div className="octo-button" style={{ color: "#000000", fontWeight: 600 }} onClick={(e) => { OctoUtils.showViewMenu(e, mutator, boardTree, pageController) }}><div className="imageDropdown"></div></div>
+							<div className="octo-button" style={{ color: "#000000", fontWeight: 600 }} onClick={(e) => { OctoUtils.showViewMenu(e, mutator, boardTree, showView) }}><div className="imageDropdown"></div></div>
 							<div className="octo-spacer"></div>
 							<div className="octo-button" onClick={(e) => { this.propertiesClicked(e) }}>Properties</div>
 							<div className={ hasFilter ? "octo-button active" : "octo-button"} onClick={(e) => { this.filterClicked(e) }}>Filter</div>
@@ -243,8 +246,7 @@ class TableComponent extends React.Component<Props, State> {
 	}
 
 	private filterClicked(e: React.MouseEvent) {
-		const { pageController } = this.props
-		pageController.showFilter(e.target as HTMLElement)
+		this.props.showFilter(e.target as HTMLElement)
 	}
 
 	private async optionsClicked(e: React.MouseEvent) {
@@ -348,7 +350,7 @@ class TableComponent extends React.Component<Props, State> {
 	async showCard(card: IBlock) {
 		console.log(`showCard: ${card.title}`)
 
-		await this.props.pageController.showCard(card)
+		await this.props.showCard(card)
 	}
 
 	focusOnCardTitle(cardId: string) {
@@ -396,13 +398,13 @@ class TableComponent extends React.Component<Props, State> {
 		if (e.keyCode === 27) {		// ESC: Clear search
 			this.searchFieldRef.current.text = ""
 			this.setState({ ...this.state, isSearching: false })
-			this.props.pageController.setSearchText(undefined)
+			this.props.setSearchText(undefined)
 			e.preventDefault()
 		}
 	}
 
 	searchChanged(text?: string) {
-		this.props.pageController.setSearchText(text)
+		this.props.setSearchText(text)
 	}
 }
